refactor: extract the alias row model to its own module
Second round of review feedback on #351.

- Revert the elective prettier pass over plans/remove-openrouter-engine.md
  and reapply only the Decision 5 amendment. Markdown read the wrapped
  "(`OPENROUTER_DEFAULT_BASE_URL` + `resolveOpenRouterApiUrl`)" as a list
  bullet and normalized the `+` to `-`, so the doc stopped saying the
  module exports both names — a content regression in the file this PR
  amends. Nothing in the repo formats markdown, so the pass was elective;
  it also de-indented a continuation line and flipped ~15 unrelated
  emphasis markers, burying the amendment in churn.

- Move AliasRow and its conversions to modelAliasRows.ts. The previous
  commit widened the component's public surface purely for a test, which
  is neither of the two patterns this repo uses; pendingLabels.ts +
  pendingLabels.test.ts is the precedent, and Board.tsx exports no helpers
  at all. The comment justifying the exports was the module boundary
  asking to be drawn. Pure move — the tests are unchanged apart from the
  import, and the round-trip guarantee they pin is identical.

- Give the "Clear it" button an aria-label. Every row renders the same
  label, so the accessible name has to carry the alias name to tell them
  apart.
